[RenderEngineVtk] Throw exceptions rather than abort() in vtkXOpenGLRenderWindow.cxx
This patch is being added to drake to bypass inconsistent error reporting
mechanisms in VTK that result in jupyter kernels restarting without any output
to the user as to what went wrong (e.g., DISPLAY is not set but they are trying
to use graphics functions).  See drake issue 18445 for more information.