[py common] Centralize some test utilities
To aid in the solution to a pydrake dependency cycle, it's convenient
for pydrake.autodiffutils and pydrake.math to simply be Python modules,
not Python packages. (A package is a module that has sub-modules.)

To that end, we ditch the `test` packages and move those modules into
the pydrake.common.test_utilities package, instead.