Rework deformable torus example to demo external force
Add PointSourceForceField as an example of a concrete ForceDensityField.

Add a `gripper` flag so that one can toggle between a parallel gripper
and a suction gripper to pick up the deformable torus.

Refactor the controller into separate files to reduce the lines of code
in the example .cc file.

Change the parallel fingers to capsule shape to avoid sticking resulting
from the sharp change in contact normal from boxes.