Sensor poses are parameters
The pose of the sensor body w.r.t. to its parent frame is now a System
Parameter and can be modified in the context. RgbdSensor and RgbdSensorAsync
change to reflect this. RgbdSensorDiscrete requires no change because it
simply exposes its underlying RgbdSensor.

This deprecates the old X_PB() method on RgbdSensorAsync in favor of the
new parameter.