Adjust tolerance for KinTrajOpt w/ new ipopt on mac arm
Resolves a CI failure at
https://drake-jenkins.csail.mit.edu/job/mac-arm-monterey-unprovisioned-clang-bazel-nightly-release/409/consoleFull

Homebrew upgraded from ipopt 3.14.14 (which passed) to ipopt 3.14.13
(which failed).  This is a new failure due to the homebrew upgrade.

The solver was actually finding a solution which passed the test, but
reporting failure because it could not achieve the desired
tolerance. Additionally, IPOPT was complaining about unbounded
variables. This PR adds bounds to the unbounded variables and adjusts
the solver convergence tolerance.