Adjust tolerance for KinTrajOpt w/ new ipopt on mac arm (#20526)
Resolves a CI failure during
mac-arm-monterey-unprovisioned-clang-bazel-nightly-release.

Homebrew upgraded from ipopt 3.14.14 (which passed) to ipopt 3.14.13
(which failed).  This is a new failure due to the homebrew upgrade.

The solver was actually finding a solution which passed the test, but
reporting failure because it could not achieve the desired
tolerance. Additionally, IPOPT was complaining about unbounded
variables. This PR adds bounds to the unbounded variables and adjusts
the solver convergence tolerance.