[workspace] Add rules_license 0.0.7
This is required by the `@platforms` external. In modern versions of
Bazel a sufficiently new version of `@rules_license` is built-in, but
the older Bazel 5.x series will fail without this. In any case, we
should affirmatively control the versions of all of our dependencies,
so relying on the built-in one is poor form anyway.