Add per-build bazelrc (#20566)
Change the CMake-generated bazelrc file to also look for an (optional)
per-build bazelrc. This gives users a way to create a bazelrc that only
applies to a specific CMake build tree, in addition to the user.bazelrc
that applies to the Drake checkout.

Also, change the Linux wheel build to use this rather than the system
bazelrc.