implements MultibodyPlant::RemoveJointActuator
In order to keep existing actuator indices valid, MultibodyTree
replaces the actuator in the list with nullptr.  As a result, we must
also introduce new supporting api: for checking if an actuator index
is valid, and for iterating over all actuator indices:
- has_joint_actuator(actuator_index)
- GetJointActuatorIndices()

It is no longer sound to iterate with a for loop of JointActuatorIndex
from 0 to num_actuators() -- both because some elements will be
omitted AND because num_actuators() is not necessarily the highest
actuator index.