Correct phong-to-pbr promotion in presence of diffuse texture (#20607)
The generic phong model uses vtkActor::Texture to define the diffuse
texture.

The Pbr illumination model uses a texture assigned to the actor's
properties (the "albedo texture"). Furthermore, that texture must
be marked as sRGB.

This correct both the assignment (when PBR is active) as well as the
upgrade when it is activated after geometries have aleady been
assigned.