Correct VTK's camera copy
We rely on setting vtkCamera's projection matrix explicitly. Previously,
it worked because we used a very simple render pipeline. However, when
upgrading to various render passes, some of those passes *clone* the
camera. The clone is incomplete and, specifically, omits our explicitly
defined projection matrix, leading to disparity between the projection
matrix we've defined and the one used in the rendering.

This patches VTK to complete the copy.