[workspace] Cull sdformat use of environment and subprocess
Actions like reading from environment variables and launching
subprocesses are important enough that Drake wants to maintain
positive control over any use of those features. Therefore, we'll
remove those classes from gz-utils so that we'll receive a build
failure in case a library upgrade starts to use them.

This turned up the fact that we've been using an $SDF_PATH environment
variable for years, without documenting it. Add a TODO to fix this at
some point.