[workspace] Cull sdformat use of environment and subprocess (#20642)
Actions like reading from the environment and launching processes are
important enough that Drake should maintain positive control over any
use of those features. Therefore, we'll remove that API from gz-utils
so that we'll receive a build failure in case a library upgrade starts
to use it.

This caught the bug that we've been using $SDF_PATH for years without
documenting it. We'll opt-out of that feature.

This is a breaking change: Drake no longer supports $SDF_PATH for
resolving SDFormat <include> statements.