Change how linux-only tests get built
It appears that if we have a .cc file that declares unit tests (e.g.,
using TEST_F), if we build it into a library against @gtest//:without_main,
the actual tests get omitted.

This tweaks the set up slightly so that if we're on linux, the test
declaration uses the .cc file directly and links with drake's default
gtest main function.

(Incidentally fixes a typo.)