Correct window sizing in RenderEngineGl (#20695)
If a RenderEngineGl is asked to render and *display* an image that didn't
match the default window size in the OpenGlContext, it would hang.

This is because the resizing would wait for xwindows to provide an expose
event. However, if the window isn't visible yet (i.e., it's being resized
before being made visible), there won't be such an event. This leads to
the process hanging, waiting for an event that will never come.

We've changed the context's default window size. The test on the
OpenGlContext is sufficient to test the expected outcome.