Implements MultibodyPlant::RemoveConstraint
Adds a pre-finalize method to remove MbP constraints.

An example use case for this is when I've parsed a gripper into my
plant with the parser, but then want to go through and replace all of
the joints with weld joints. In that workflow, I'll also have to
remove any constraints that were added by the parser, but will now
become invalid.

Also adds an error message when somebody adds constraints and
_then_ calls set_discrete_contact_approximation to TAMSI.