MJCF parser supports eulerseq (and fixes the default) (#20620)
Towards #20232.  Related to #20444.

In #20232 we realized that the default eulerseq in MJCF is "xyz",
which is NOT the standard eulerseq in urdf/sdf. This PR fixes a bug in
which we assumed the wrong eulerseq by default, leading to incorrect
kinematics in some parsed models. It also adds support for non-default
eulerseq.

It also adds a test so that we can readily visualize the parsed
kinematics from the dm_control suite which are being used to test the
parser.