fixup! Get vtktoken working
This module is a c-c-combo breaker. It's first-party code but doesn't
live with the other first-party stuff like kwsys, but rather in ThirdParty.
However, it doesn't build like any other ThirdParty code -- it doesn't
have a trampoline header that adjusts the include paths accordingly. We
need to work around that with a new settings option to tack on to the
include path.