Build macOS wheels for multiple Python versions (#20616)
Rework macOS wheel build scripts to use Python-version-specific
permanent locations for all operations, with temporary symlinks used for
operations shared with the Linux wheel builds that expect non-versioned
locations. Also rework builder to support a set of Python targets and to
be able to build multiple wheels in a manner similar to how things work
for Linux.

This changes the macOS wheel build to produce wheels for (as of writing)
both Python 3.11 and 3.12.