[geometry] Serve the environment map over http
The behavioral change here is a small one: when setting an environment
map instead of it being copied into the property message as an inline
data URI, the property message only contains the URL and image data is
served over http.

Primarily, this allows the browser to avoid re-requesting the image in
case it's already downloaded it previously. It also has side benefits:
the http image becomes easier to debug -- in the developer console it
shows up as a separate request, including an inline image preview.  In
theory it should also be faster to skip the base64 encoding in C++ and
decoding in the browser, though that's not really measurable for just
the environment map.

The bigger picture is that the CAS technology introduced here can also
eventually be used for resources beyond just the environment map, such
as texture files, geometry files, object mesh files, etc.

Forr testing the changes to meshcat.cc, the pre-existing unit and
manual tests already have sufficient coverage.

XXX The meshcat_manual_test can lie now, pulling from disk cache.

XXX Figure out where to call ShrinkToFit.