[workspace] Avoid crashes in python repository rule
Our intent is that users can BYO Python interpreter. However, the dict
access into _VERSION_SUPPORT_MATRIX would crash when not running on a
supported platform, even if the user had specified a valid interpreter
in the rule attribute.

We support enough different Python versions and platforms now (e.g.,
wheel builds) that trying to keep this warning up-to-date is no longer
worth it, so remove it and simplify the nearby code.