[tools] Change how wheel builds opt-out of LCM java
Instead of using the "detected OS" ("manylinux" or "macos_wheel") to
implicitly opt-out (by virtue of it not being "ubuntu"), we now have
an (internal use only) opt-out flag that wheel builds toggle acutuely.

This still isn't terribly elegant, but it's at least slightly more
readable and documented. The important upside is removing another use
of the determine_os() function, which is scheduled for removal.