[workspace] Remove macOS-specific logic for zlib (#20829)
The benefit of teaching Bazel exactly which zlib headers XCode will
use on macOS isn't worth this much complexity. We're trying to reduce
our use of the determine_os() logic to better support more platforms.
Rather than rewriting this logic to be safe on additional platforms,
it's easier just to toss it.