Fix bug in glTF depth and label representations (#20854)
When glTF objects were rendered in depth or label images, they would be
(at least) rotated from where they should be. In fact, if a non-unit scale
factor was applied, they'd also be scaled too much (the scale factor
squared).

This is because the creation of the color representation would modify the
reference version before the depth and label versions were created. So,
those later representations would accumulate an extra concatenation of the
T_GF transform (the one that related the glTF file frame with the Drake
geometry frame -- this transform incorporated both the 90-degree rotation
from y-up to z-up and the Mesh/Convex scale factor).