Teach new_release to use upgrade scripts
Teach new_release to be able to upgrade things that need an external
script to perform the update. Right now this is crate_universe. Add
crate_universe to a cohort with clarabel_cpp_internal. Hack metadata.py
to supply metadata for a 'fake' crate_universe, since Bazel-side breaks
it down into individual crates.