[tools] Finish writing lcm_gen for C++
Features of Drake's new lcm_gen:
- Portable:
  - Single file, written in Python 3.
  - Depends only on the Python 3 standard library.
- Easy to modify:
  - No compilation step.
  - High level language.
- Improved C++ output:
  - C++17 mode.
  - Generated code has no dependencies other than the standard library.

This version only supports C++ codegen, because the upstream codegen is so bad.
We could imagine adding Python and/or Java codegen later, if we need it.

This commit introduces lcm_gen support for strings and arrays.
This does not yet switch Drake to use the new tool by default.