Don't install as build step
Rip out use of ExternalProject to drive the build, as we aren't doing
anything that really benefits from it. Move the install logic to
install(CODE) so that it happens via 'make install' (or the equivalent),
rather than as part of the build (e.g. 'make'). Similarly, move the call
that announces build information to the configure stage.