Simplify RenderEngineVtk tests (#20701)
There are a *mountain* of tests which poke at narrow aspects of
the render engine's rendering capability. It leads to a huge
number of renderings being made.

This replaces the host with a couple of renderings by populating a
scene with many elements, each of which contribute to a final image.
By looking at the image, we can assess whether the rendering is
correct. And by having the unit tests re-render and compare against
that reference, we can cover a multitude of features with a single
rendering.

This includes generating a new obj at build time: box_checkered.obj. It
applies a checkered texture to the box by swapping material files.