[setup] Add bootstrap support for Noble
This get us to the point of seeing the compilation errors. We don't
have CI for this yet, so this is developed & tested under docker.

The majority of the changes are to our allow-lists for OS versions,
and tweaking some version numbers in the apt lists.

We drop valgrind-dbg because it doesn't exist anymore; instead, it
uses the modern Debian non-package mechanism to fetch debug symbols.

We drop kcov because it's no longer in Ubuntu. Hrm.