Fix lights expressed in camera frame for RenderEngineVtk (#20930)
* Fixes VTK illumination bugs

See the documentation in fix_illumination_bugs.patch for details. But this
corrects:

1. Missing tests on the "positionality" of a light,
2. Computation using uninitialized memory.

* Fix lights expressed in camera frame for RenderEngineVtk

The old code misrepresented the device frame; it was incomplete. It flipped
to left-handed when it is more accurate to rotate the around Cx 180 deg.

This also changes the light direction vector semantics. The input value
need not be unit length, just normalizable. RenderEngineVtk had to change
what it was doing, but RenderEngineGl was already doing this. Only its
test was expanded to explicitly cover normalization.