[mesh_to_model] Warn about implausible scale
Wavefront .obj files do not specify or store units for the vertex
vectors, and different tools use different units.

It is impractical and imprecise to try to infer units from comments that
might be in the input file (though some software tries):
https://stackoverflow.com/questions/47995531/three-js-load-proper-model-scale-size

Instead, this patch checks the volume of the output against implausibly
large and small volumes, and issues warnings.