Fixup InverseDynamics to use the actuation matrix.
Clarifies that the output of the inverse dynamics controller is
actuation input, not generalized forces, and uses the actuation matrix
to achieve this. Previously, by connecting the output of the
InverseDynamics system to an actuation input port, we implicitly
assumed that the actuator matrix was the identity matrix. If the
actuators were in a different order than the joints (generalized
forces), then our method produced the wrong actuator inputs.