Drake-generated meshes render faceted in meshcat
When Meshcat sends a Drake-generated mesh (as indicated by a call to
a SetTriangleMesh() variant), that mesh's material is configured to render
the mesh faceted.

The default behavior of smoothing them is unhelpful. Meshes of strongly
angular objects (e.g., boxes) have their creases smoothed over. It can be
difficult to fully understand the topology of the generated mesh with
smoothing.