Add new algorithm, IrisInConfigurationSpaceFromCliqueCover, for efficiently placing IRIS regions.
Also adds a new subdirectory planning/iris

This algorithm attempts to cover a large portion of the free configuration space in as few IRIS  
regions as possible. It builds a visibility graph, constructs a small clique cover of the visibility   
graph, and uses the resulting cliques to position and direct IRIS region growth. 

Note: This does not change the existing IRIS implementation.