Implements DerivativeTrajectory.
Trajectory objects provide derivatives by implementing `DoEvalDerivative`
_and_ `DoMakeDerivative`. `DoEvalDerivative` evaluates the derivative value at
a point in time. `DoMakeDerivative` returns a new Trajectory object which
represents the derivative.

In some cases, it is easy to implement `DoEvalDerivative`, but difficult or
inefficient to implement `DoMakeDerivative` natively. And it may be just as
efficient to use `DoEvalDerivative` even in repeated evaluations of the
derivative. The %DerivativeTrajectory class helps with this case -- given a
`nominal` Trajectory, it provides a Trajectory interface that calls
`nominal.EvalDerivative()` to implement `Trajectory::value()`.

Resolves #20982.