Change mechanism for proximity geometry transparency (#21015)
When using VisualizationConfig, the default color for proximity geometries
was 100% red with 50% alpha.

In rendering, it's generally a bad idea to define albedo (aka diffuse)
color at 100%, and bad for red particularly. This has been backed off to
80% red.

By defining the *color* to be 50% transparent, the alpha slider for
proximity can no longer make it opaque.

This changes the default configuration to have the same default behavior
(red proximity geometries with 50% transparency), but defaults the
*color* to be opaque and configures the slider to be at 50% so full opacity
is now available.

The old behavior can be achieved by changing the default_proximity_color
back to [1, 0, 0, 0.5] and the initial_proximity_alpha to 1.