[render_vtk] Obey extensionsUsed and extensionsRequired
When given a glTF with an unsupported extensionsRequired, throw an
error. (Previously, we would print an error to stderr but bash on
regardless.)

When given a glTF with an unsupported extensionsUsed (not required),
don't report errors about known image formats (e.g., ktx2), rather
only about images we need but cannot actually load. The warning about
the unsupported texture extension is sufficient on its own.