[install] Deprecate drake_bazel_installed support
This doesn't work with bzlmod, and I can't figure out how to make it
work. Even with repository rules, every few months it breaks and I
can't invest the time anymore to keep it working.

If there are Bazel users out there who want this to work and can
figure out how to make it happen, we'd certainly welcome a pull
request with an improved implementation.

For Bazel users who only need pydrake, note that the rules_python
support for pip wheel is pretty good these days; using the pydrake
wheels via rules_python is a good option.

During the deprecation window we will continue to install the code in
case it still happens to work for you, but we are no longer testing it
in our CI so it will probably bitrot at some point.