[tools] Delete troublesome compiler rules (#21115)
These rules are painful to maintain across Bazel upgrades.

Remove the drake_assert_compile tests. They haven't caught any bug in
a long time.

Remove the print_host_settings utility. It was supposed to be helpful
for bug reporting, but nobody has used it in the last ~4 years and
it's not part of our reporting template anymore.