Push convex hull down into hydroelastic and deformable geometries
When a Convex shape is used, the hydroelastic and deformable representations
should always use its convex hull. This updates the implementations to
make this so.

This also adds a feature that specifying a Mesh to be hydroelastic
compliant (but using, for example, an obj instead of a volume vtk) allows
us to create a compliant representation based on its convex hull.

TODO:
 - When #21124 merges, rebase on that and clean up the changes to
   deformable_contact_geometries.h.