MakeConvexHull now supports glTF files (#21131)
We're using VTK to parse the glTF file. For VTK's parser to work, it needs
the OpenGl2 rendering artifacts initialized. This required moving VTK
module initialization out of `internal_render_engine_vtk.cc` and into its
own compilation unit to be shared.

It exercises the API in planar_scenegraph_visualizer.

While the API could probably be promoted to part of the stable API, that
decision is being deferred for now. It is bound as _MakeConvexHull.