Fix rigid geometry registration for deformable contact (#21124)
When deformable bodies are present, we attempt to assign a mesh
representation to all rigid geometries with a resolution hint so that
they can interact with deformable geometries. Failing that, a geometry
should not throw immediately throw but log a one-time warning instead
and subsequently ignore interactions with deformable bodies.

This has always been the design intention, but previously the
implementation incorrectly aborts in the (rare) presence of a half space
with resolution hint property specified.