Overhaul how VERSION.TXT is generated
Generate VERSION.TXT at build time rather than configure time. Since
configuring only needs to happen infrequently, it was possible that the
contents would be significantly out of date if a user configures, then
after some time, fetches the latest changes and builds again. By moving
generation to build time, we ensure that the time stamp (which includes
a time and not just a day) is always maximally correct, and likewise
that the git SHA is correct.

Additionally, two other changes are made to the VERSION.TXT contents.
First, rather than just a time stamp, the default value is now an actual
version identifier following the usual pattern of experimental builds
(i.e. 0.0.YYYYMMDD.HHMMSS+git<sha8>). Second, if the environment
variable DRAKE_VERSION is set, its value overrides the value that would
otherwise appears in VERSION.TXT. This mechanism is intended to allow
official builds a way to inject release version numbers, as opposed to
the current mechanism of replacing VERSION.TXT.

Finally, ensure that wheel builds export the wheel version to the build
environment so that it is picked up when generating VERSION.TXT.