[multibody] AddRigidBody offers a default inertia
We want to enable MbP SAP by default, but if the user welded bodies
with NaN inertia the SAP-flavored MbP will throw an exception (whereas
previously this was allowed with a continuous-time or TAMSI plant).

Instead of trying to adjust SAP to somehow work around the NaNs, we've
decided that we must get rid of them in the first place. That means
that now the default constructor for SpatialInertia is footgun, where
our users who were using it as the most compact syntax when adding
welded bodies will now start to crash with SAP.

We attack this from two sides:

- MbP::AddRigidBody now offers a default value for the inertia. When
  users are adding welded bodies where they don't care about it, the
  new default makes it easy and compact to add the body.

- Deprecate the footgun SpatialInertia constructor, and in partibular
  use its deprecation message to teach users about the new AddRigidBody
  suggestion.