Clarify some contracts relating to System::SetDefaultContext()
System::SetDefaultContext() calls SetDefaultState() and
SetDefaultParameters() by passing a const Context& and a mutable State or
Parameters. However, that mutable output parameter was actually a view
into the const Context.

This undocumented relationship between the two parameters is dangerous.
The relationship should either be documented or ended. Here we opt to
document that the user should not assume the input parameter and output
parameter are independent (without promising the aliasing currently taking
place).

Furthermore, SetDefaultContext() doesn't document the order in which state
and parameters are set. In this case, we document that the default
parameter values have been set prior to setting the default state (and
provide regression tests). (We've also re-ordered the definition/declaration
of the two functions to mirror that invocation ordering).