System::SetDefaultState() & SetDefaultParameters() clarified and tested
System::CreateDefaultContext would call SetDefaultState() and
SetDefaultParameters() by passing a const Context& and a mutable State or
Parameters. However, that mutable parameter was actually a view into the
const Context. The fact that they are mutable views into a const parameter
should not go undocumented. (It should either *not* be a mutable view
or should be documented.) This opts for the latter, documenting the
relationship and providing a regression test.