Overhaul how VERSION.TXT is generated
Generate VERSION.TXT at build time rather than configure time. Since
configuring only needs to happen infrequently, it was possible that the
contents would be significantly out of date if a user configures, then
after some time, fetches the latest changes and builds again. By moving
generation to build time, we ensure that the time stamp (which includes
a time and not just a day) is always maximally correct, and likewise
that the git SHA is correct.

Additionally, rather than just a time stamp, the first of the two values
in VERSION.TXT now defaults to an actual version identifier of the usual
form for experimental builds (i.e. 0.0.YYYYMMDD.HHMMSS+git<sha8>). This
can be overridden by undocumented CMake variables, which are used by
wheel and CI builds to inject "real" version numbers. (The git SHA can
be similarly overridden, which is needed for Linux wheel builds as the
Docker build environment has only the raw sources and is therefore
unable to obtain the git SHA on its own.)

Finally, tweak wheel builds to make use of these new mechanisms.