Allow model directive 'add_model' to pose free body w.r.t arbitrary frame
Previously, when specifying `default_free_body_pose` in a dmd.yaml file,
the `base_frame` attribute of each Transform had to be omitted (and was
always interpreted as "world").

Now the `base_frame` attribute can be set. The options are:

 - No behavioral change if `base_frame` is omitted.
 - Set to "world" (the same as omitting it).
 - Set to the name of a *frame* or *body* that has already been added
   (including by the current `add_model` directive).
   - The specified name *must* be a scoped name.
   - the pose will be interpreted as being measured and expressed in the
     `base_frame` frame.
   - The posed body is *not* considered a "free body" by Drake;
     body.is_floating() will return `false` and
     MbP::Get/SetDefaultFreeBodyPose() will have no meaningful effect.

We also update the documentation and and opportunistic document typo clean
up.