Fixes bugs in IrisInConfigurationSpaceFromCliqueCover (#21237)
The function IrisInConfigurationSpaceFromCliqueCover had ill defined behavior when a clique with a degenerate MinimumVolumeCircumscribedEllipsoid is encountered. The corresponding IrisWorker would get stuck attempting to build the set from the degenerate clique and repeatedly triggering the same runtime error without discarding the clique after one try. The fix is to pop the clique from the queue when the circumscribed ellipsoid fails. This PR further fixes the minimum_clique_size override.