Meldis shows convex hull for Mesh shapes with the proximity role (#21281)
DrakeVisualizer now takes responsibility for accounting for the *role* in
dispatching Mesh-based meshes.

Indirectly, we also tweak a message in meldis to indicate which visualization
"channel" is ignoring a load message. It has been mildly confusing to
see multiple instances of "Ignoring load message" appear while the scene
nevertheless updates. The addition of the channel name helps clarify the
messages.

To test:

1. Launch meldis
2. bazel run //tools:model_visualizer_private -- package://drake/examples/scene_graph/cuboctahedron_with_hole.obj
3. Make sure you open the browser to the meldis URL.

Run in master, what is shown when "Collision Geometry" is the textured
cuboctahedron. In this branch, it is replaced with the red convex hull.