Make SAP the default discrete contact solver.
This PR changes the default discrete solver from TAMSI to SAP, using the kSimilar contact approximation.
The update essentially resolves to making  MultibodyPlantConfig::discrete_contact_approximation="similar".

We mark this as a breaking change, since changing the underlying contact approximation can lead to slightly different simulation results.
This might not be noticeable in most cases, but users should at least be aware of this.
In particular, most users of SAP ommit setting disspation values and use the defult value of relaxation_time, which is non-zero.
However, the newly adopted approximation uses the more physical (and desirable) Hunt & Crossley model, with a default value of dissipation equal to zero.
Therefore users must be aware that from now on, modeling inelastic collisions requires explicitly setting a value for the hunt and crossley dissipation, drake:hunt_crossley_dissipation when setting proximity properties through a model file.