Make SAP the default discrete contact solver.
This PR changes the default discrete solver from TAMSI to SAP, using the
kLagged convex approximation of TAMSI's contact model. The update
essentially resolves to making
MultibodyPlantConfig::discrete_contact_approximation="lagged".

Though this is not a breaking change, it will definitely cause different
simulation results, even if more correct. This might not be noticeable
in most cases, but users should at least be aware of this.

There is not much of a difference in terms of model files for those
users previously using kTamsi, since kSimilar simply is a convex
approximation of the same TAMSI's  models. For those users that have
been using kSap, the model of dissipation changes from a linear
Kelvin-Voigt model controlled by relaxation_time to Hunt & Crossley,
controlled by hunt_crossley_dissipation.