[tools] Prepare for Noble
Fix python version in setup scripts; since the last time we prepped
for Noble, the default Python switched from 3.11 to 3.12.

Add several GCC 13 error suppressions.  We'll need to circle back and
deal with these later, but for now it's better to silence them so we
can focus on any other build problems.

Override nlohmann/json spaceship probing. It's broken on GCC 13.