[tools] Fix kcoverage vs jupyter runfiles (#21350)
When testing a Jupyter notebook, the `exe = os.path.realpath(exe)`
adjustment in kcoverage.py defeats the ability of the codegen'd main
program to find its runfiles.

We work around this by moving the runfiles resolution from the
generated main to the library main. Bazel's py_binary trampoline
loads the main library from the proper runfiles path, which
allows runfiles to be found again.

This is a hotfix for f472c414.