Fix docs and overloading
Hoist object-animation-timing doc to class overview; the caveats are
going to apply to a whole bunch of functions, so we should explain it
in one place and xref from each function.

Remove timed-untimed overloading. The timing must be a variable, not a
static dispatch. Some call sites will themselves be given optional
time and need to forward it along, i.e., timed-vs-untimed is not
always statically known at 100% of callers.

Remove "experimental" marking. We agree it's not necessary.