Deprecate --version in repack_deb.py
After recent changes, VERSION.TXT should always contain a meaningful and
correct version number. Accordingly, we a) should stop assuming it is an
unprefixed timestamp, and b) no longer need the ability to override
VERSION.TXT, which allows for some modest simplification.

For now, the option is retained but ignored, because we will first need
to change CI to stop passing it, after which it can be removed entirely.