[geometry] Factor recording implementation out of Meshcat
The user-facing API is unchanged.

This provides for better encapsulation, less brittleness, and easier
unit testing -- all of which will be critically important as we start
to add support for timed SetObject calls.

This also fixes an implementation inconsistency where the *data type*
of the property being set affected whether or not during a recording
with set_visualizations_in_recording=false an untimed SetProperty
would be shown live or not.