[manipulation] Purge some obsolete tools (#21398)
The meshlab inertia thing is obsolete now that we have mesh_to_model.

The convert_to_visualize_collisions is obsolete now that we have
proximity visualization available by default.

Add a purge reminder to the show_model shim.

The important outcome here is that we can delete some __init__ files,
which cause path-confusion trouble for bzlmod.