[examples] Clean up visibility and paths (#21401)
The default visibility (for rules that aren't otherwise marked) shall
always be private; making example code or models public should always
be opt-in.

Align opt-in in to public visibility to exactly match things that are
installed; if its not installed, then it shouldn't be used outside of
Drake.

Avoid "subpackages" spelling for visibility; for clarity we should
either make things fully public (installed) or else limit to an
allow-list of specific other packages one by one.

Now that all examples code is suitably private, re-spell the imports
to lose the "drake" part of the module namespace. The old spelling of
imports is dispreferred and will not work with upcoming bzlmod.