DeformableModel allows scalar conversion (#21409)
Cloning to double is always allowed. Cloning to non-double scalar types
is only allowed when the DeformableModel is empty.

Move the deformable configuration port accessor from DeformableModel to
MultibodyPlant to help facilitate AddMultibodyPlantSceneGraph in the
future. In particular, the port is allocated pre-finalize now.

Instead of offering a pointer to the owning MbP at
DeclareSystemResources, PhysicalModel now takes a mutable pointer to the
prefinalized owning pointer at construction and nulls the pointer at
DeclareSystemResources.

Change DummyModel into DummyPhysicalModel.